Move api mapping methods to api controller
