+++ ngrinder-controller/src/main/java/org/ngrinder/user/controller/UserApiController.java
@@ -4,10 +4,15 @@ import static java.util.stream.Collectors.toList;
 import static org.ngrinder.common.util.CollectionUtils.buildMap;
 import static org.ngrinder.common.util.CollectionUtils.newArrayList;
 import static org.ngrinder.common.util.ObjectUtils.defaultIfNull;
-import static org.ngrinder.common.util.Preconditions.checkNotEmpty;
+import static org.ngrinder.common.util.Preconditions.*;
+import static org.ngrinder.common.util.Preconditions.checkNull;
 
 import com.google.common.collect.Lists;
+import com.google.gson.annotations.Expose;
+import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.collections.Predicate;
 import org.apache.commons.lang.StringUtils;
+import org.ngrinder.common.constant.ControllerConstants;
 import org.ngrinder.common.controller.BaseController;
 import org.ngrinder.common.controller.RestAPI;
 import org.ngrinder.infra.config.Config;
@@ -104,7 +109,7 @@ public class UserApiController extends BaseController {
 	 */
 	@GetMapping("/{userId}/detail")
 	@PreAuthorize("hasAnyRole('A')")
-	public Map<String, Object> getOne(@PathVariable final String userId) {
+	public Map<String, Object> getOneDetail(@PathVariable final String userId) {
 		User one = userService.getOneWithFollowers(userId);
 		Map<String, Object> model = buildMap(
 			"user", one,
@@ -156,21 +161,227 @@ public class UserApiController extends BaseController {
 		user.setFollowers(null);
 		user.setOwners(null);
 
-		model.put("followers", followers.stream().map(UserController.UserSearchResult::new).collect(toList()));
-		model.put("owners", owners.stream().map(UserController.UserSearchResult::new).collect(toList()));
+		model.put("followers", followers.stream().map(UserSearchResult::new).collect(toList()));
+		model.put("owners", owners.stream().map(UserSearchResult::new).collect(toList()));
 		model.put("allowShareChange", true);
 		model.put("userSecurityEnabled", config.isUserSecurityEnabled());
 	}
 
-	private List<UserController.UserSearchResult> getSwitchableUsers(User user, String keywords) {
+	private List<UserSearchResult> getSwitchableUsers(User user, String keywords) {
 		if (user.getRole().hasPermission(Permission.SWITCH_TO_ANYONE)) {
-			List<UserController.UserSearchResult> result = newArrayList();
+			List<UserSearchResult> result = newArrayList();
 			for (User each : userService.getPagedAll(keywords, PageRequest.of(0, 10))) {
-				result.add(new UserController.UserSearchResult(each));
+				result.add(new UserSearchResult(each));
 			}
 			return result;
 		} else {
 			return userService.getSharedUser(user);
+		}
+	}
+
+
+	/**
+	 * Save or Update user detail info.
+	 *
+	 * @param user        current user
+	 * @param updatedUser user to be updated.
+	 * @return "redirect:/user/list" if current user change his info, otherwise return "redirect:/"
+	 */
+	@ResponseBody
+	@RequestMapping("/save")
+	@PreAuthorize("hasAnyRole('A') or #user.id == #updatedUser.id")
+	public String save(User user, @RequestBody User updatedUser) {
+		checkArgument(updatedUser.validate());
+		if (user.getRole() == Role.USER) {
+			// General user can not change their role.
+			User updatedUserInDb = userService.getOne(updatedUser.getUserId());
+			checkNotNull(updatedUserInDb);
+			updatedUser.setRole(updatedUserInDb.getRole());
+
+			// prevent user to modify with other user id
+			checkArgument(updatedUserInDb.getId().equals(updatedUser.getId()), "Illegal request to update user:%s",
+				updatedUser);
+		}
+		save(updatedUser);
+		return returnSuccess();
+	}
+
+	private User save(User user) {
+		if (StringUtils.isBlank(user.getPassword())) {
+			return userService.saveWithoutPasswordEncoding(user);
+		} else {
+			return userService.save(user);
+		}
+	}
+
+
+	@PreAuthorize("hasAnyRole('A')")
+	@GetMapping({"/role", "/role/"})
+	@ResponseBody
+	public EnumSet<Role> roleSet() {
+		return EnumSet.allOf(Role.class);
+	}
+
+	/**
+	 * Delete users.
+	 *
+	 * @param userIds comma separated user ids.
+	 * @return "redirect:/user/"
+	 */
+	@PreAuthorize("hasAnyRole('A')")
+	@DeleteMapping({"", "/"})
+	public HttpEntity<String> deleteUsers(User user, @RequestParam String userIds) {
+		String[] ids = userIds.split(",");
+		for (String eachId : ids) {
+			if (!user.getUserId().equals(eachId)) {
+				userService.delete(eachId);
+			}
+		}
+		return successJsonHttpEntity();
+	}
+
+	/**
+	 * Check if the given user id already exists.
+	 *
+	 * @param userId userId to be checked
+	 * @return success json if true.
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping("/{userId}/check_duplication")
+	public HttpEntity<String> checkDuplication(@PathVariable String userId) {
+		User user = userService.getOne(userId);
+		return (user == null) ? successJsonHttpEntity() : errorJsonHttpEntity();
+	}
+
+	/**
+	 * Get users by the given role.
+	 *
+	 * @param role user role
+	 * @return json message
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping(value = {"/", ""}, method = RequestMethod.GET)
+	public HttpEntity<String> getAll(Role role) {
+		return toJsonHttpEntity(userService.getAll(role));
+	}
+
+	/**
+	 * Get the user by the given user id.
+	 *
+	 * @param userId user id
+	 * @return json message
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping(value = "/{userId}", method = RequestMethod.GET)
+	public HttpEntity<String> getOne(@PathVariable("userId") String userId) {
+		return toJsonHttpEntity(userService.getOne(userId));
+	}
+
+	/**
+	 * Create an user.
+	 *
+	 * @param newUser new user
+	 * @return json message
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping(value = {"/", ""}, method = RequestMethod.POST)
+	public HttpEntity<String> create(@ModelAttribute("user") User newUser) {
+		checkNull(newUser.getId(), "User DB ID should be null");
+		return toJsonHttpEntity(save(newUser));
+	}
+
+	/**
+	 * Update the user.
+	 *
+	 * @param userId user id
+	 * @param update update user
+	 * @return json message
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping(value = "/{userId}", method = RequestMethod.PUT)
+	public HttpEntity<String> update(@PathVariable("userId") String userId, User update) {
+		update.setUserId(userId);
+		checkNull(update.getId(), "User DB ID should be null");
+		return toJsonHttpEntity(save(update));
+	}
+
+	/**
+	 * Delete the user by the given userId.
+	 *
+	 * @param userId user id
+	 * @return json message
+	 */
+	@RestAPI
+	@PreAuthorize("hasAnyRole('A')")
+	@RequestMapping(value = "/{userId}", method = RequestMethod.DELETE)
+	public HttpEntity<String> delete(User user, @PathVariable("userId") String userId) {
+		if (!user.getUserId().equals(userId)) {
+			userService.delete(userId);
+		}
+		return successJsonHttpEntity();
+	}
+
+	/**
+	 * Search user list on the given keyword.
+	 *
+	 * @param pageable page info
+	 * @param keywords search keyword.
+	 * @return json message
+	 */
+	@RestAPI
+	@RequestMapping(value = "/search", method = RequestMethod.GET)
+	public HttpEntity<String> search(User user, @PageableDefault Pageable pageable,
+									 @RequestParam(required = true) String keywords) {
+		pageable = new PageRequest(pageable.getPageNumber(), pageable.getPageSize(),
+			defaultIfNull(pageable.getSort(),
+				new Sort(Sort.Direction.ASC, "userName")));
+		Page<User> pagedUser = userService.getPagedAll(keywords, pageable);
+		List<UserSearchResult> result = newArrayList();
+		for (User each : pagedUser) {
+			result.add(new UserSearchResult(each));
+		}
+
+		final String currentUserId = user.getUserId();
+		CollectionUtils.filter(result, new Predicate() {
+			@Override
+			public boolean evaluate(Object object) {
+				UserSearchResult each = (UserSearchResult) object;
+				return !(each.getId().equals(currentUserId) || each.getId().equals(ControllerConstants.NGRINDER_INITIAL_ADMIN_USERID));
+			}
+		});
+
+		return toJsonHttpEntity(result);
+	}
+
+	public static class UserSearchResult {
+		@Expose
+		final private String id;
+
+		@Expose
+		final private String text;
+
+		public UserSearchResult(User user) {
+			id = user.getUserId();
+			final String email = user.getEmail();
+			final String userName = user.getUserName();
+			if (StringUtils.isEmpty(email)) {
+				this.text = userName + " (" + id + ")";
+			} else {
+				this.text = userName + " (" + email + " / " + id + ")";
+			}
+		}
+
+		public String getText() {
+			return text;
+		}
+
+		public String getId() {
+			return id;
 		}
 	}
 }
